Move MatchTypeResponse into a separate file
